requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Deprecated==1.2.14
distro==1.9.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
icecream==2.1.3
idna==3.6
inflection==0.5.1
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.2.1
pyairtable==2.3.0.post1
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.1
